Fix use of norm and remove vecnorm
The behavior of `norm(x, p)` for matrices changed in Julia PR 27401 to
be the same as `norm(vec(x), p)` and `vecnorm` was removed. The previous
behavior of `norm` for matrices is now captured by `opnorm`. This
updates our extensions to these functions to match, and emits
deprecation warnings as appropriate.

As a somewhat related change, this also deprecates the positional
dimension argument to `sum` in favor of keyword arguments, in keeping
with Base.