Load solver packages using Requires.jl
Currently the package implements its own version of optional
dependencies which checks the current user's package installation
directories for the given solver packages. This has issues both with
precompilation and with Pkg in Julia 0.7 and higher.

Now, historically the Requires package has been a hack, but a recent
overhaul makes it more legit to use. It is currently the best, if only,
solution to optional dependencies.

This change moves all solver selection logic into Requires. This should
be largely user-invisible, with the very notable exception that the
first matching solver package is NOT automatically loaded; the user will
now need to explicitly load a solver package in order to use Convex.
The order of `using Convex` and `using SomeSolver` does not matter,
however.