Fix method overwrite warning when loading the package
This was caused by providing a default value of `nothing` for the
`model` argument in the `Problem` constructor, which ends up creating a
two-argument method that is also created further down by virtue of using
default values.

By simply removing the default value for `model`, we remain at feature
parity with the methods prior to the change and we don't overwrite
anything.