Make ConicObj its own type and use OrderedCollections
This replaces the dependency on DataStructures, which is quite large,
with one on OrderedCollections. This is a lighter weight package and
we're only using one piece of it.

The meat of this change is making ConicObj its own type. This reduces
the amount of type piracy in the package. Currently the package defines
multiplication of a sparse matrix and an ordered dictionary, as one
example.